+++ tests/unit/ResponseTest.php
@@ -1,0 +1,11 @@
+<?php
+
+// class ResponseTest extends \PHPUnit\Framework\TestCase
+// {
+//   public function testResponse()
+//   {
+//     $res = \KenFramework\Response::resp(true, "Generic Response", 201);
+//     $this->assertTrue($res['success'], "Generic Response was not successful");
+//     $this->assertEquals("Generic Response", $res['message'], "Generic Response Message Failed");
+//   }
+// }