+++ doc/changelog.md
@@ -1,10 +1,23 @@
 # Change Log
 
+## v0.6.2
+
+### Added clarifying comments
+
+- Comments are not entirely done but more have been added and updated
+- The model class can now accept null or an empty array for the \$data parameter and will not throw an error
+
+## v0.6.1
+
+### Changes to binding parameters function
+
+- Fixed a bug where routes that don't have params weren't run
+
 ## v0.6.0
 
 ### Massive changes to the structure of the library
 
-- Now using PSR-4 autoloading. No need for a massible include file.
+- Now using PSR-4 autoloading. No need for a massive include file.
 - Dispatcher class renamed to model class. Intended to be extended by features to create models.
 - Controller class now intended to be extended to describe feature controllers.
 - Ken core will be its own project and files have been removed.